requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
pymongo==4.6.0
PyJWT==2.8.0
bcrypt==4.1.2
Flask-Limiter==3.5.0
dnspython==2.4.2

